+++ tools/integration_tests/util/operations/dir_operations.go
@@ -26,12 +26,6 @@ import (
 	"strconv"
 	"strings"
 	"testing"
-
-	"cloud.google.com/go/storage"
-	control "cloud.google.com/go/storage/control/apiv2"
-	"cloud.google.com/go/storage/control/apiv2/controlpb"
-	"github.com/googlecloudplatform/gcsfuse/v2/tools/integration_tests/util/client"
-	"golang.org/x/net/context"
 )
 
 const FilePermission_0400 = 0400
@@ -164,36 +158,4 @@ func DirSizeMiB(dirPath string) (dirSizeMB int64, err error) {
 	})
 	dirSizeMB = totalSize / MiB
 	return dirSizeMB, err
-}
-
-func DeleteManagedFoldersInBucket(ctx context.Context, client *control.StorageControlClient, managedFolderPath, bucket string) {
-	//gcloudDeleteManagedFolderCmd := fmt.Sprintf("alpha storage rm -r gs://%s/%s", bucket, managedFolderPath)
-	folderPath := fmt.Sprintf("projects/_/buckets/%v/managedFolders/%v/", bucket, managedFolderPath)
-	req := &controlpb.DeleteManagedFolderRequest{
-		Name: folderPath,
-	}
-	if err := client.DeleteManagedFolder(ctx, req); err != nil && !strings.Contains(err.Error(), "The following URLs matched no objects or files") {
-		log.Fatalf(fmt.Sprintf("Error while deleting managed folder: %v", err))
-	}
-}
-
-func CreateManagedFoldersInBucket(ctx context.Context, client *control.StorageControlClient, managedFolderPath, bucket string) {
-	//gcloudCreateManagedFolderCmd := fmt.Sprintf("alpha storage managed-folders create gs://%s/%s", bucket, managedFolderPath)
-	mf := &controlpb.ManagedFolder{}
-	req := &controlpb.CreateManagedFolderRequest{
-		Parent:          fmt.Sprintf("projects/_/buckets/%v", bucket),
-		ManagedFolder:   mf,
-		ManagedFolderId: managedFolderPath,
-	}
-	if _, err := client.CreateManagedFolder(ctx, req); err != nil && !strings.Contains(err.Error(), "The specified managed folder already exists") {
-		log.Fatalf(fmt.Sprintf("Error while creating managed folder: %v", err))
-	}
-}
-
-func CopyFileInBucket(ctx context.Context, storageClient *storage.Client, srcfilePath, destFilePath, bucket string, t *testing.T) {
-	//gcloudCopyFileCmd := fmt.Sprintf("alpha storage cp %s gs://%s/%s/", srcfilePath, bucket, destFilePath)
-	err := client.UploadGcsObject(ctx, storageClient, srcfilePath, bucket, destFilePath, false)
-	if err != nil {
-		t.Fatalf("Error while copying file : %w", err)
-	}
 }
